pyproject.toml
[project]
name = "homepage"
version = "0.1.0"
description = "Personal homepage generator"
requires-python = ">=3.11"
dependencies = [
    "markdown==3.8.0",
    "python-markdown-math==0.9",
    "pyyaml==6.0.2",
    "jinja2==3.1.6",
    "watchdog==6.0.0",
    "rcssmin>=1.1.2",
]
